@@ -1,51 +1,104 @@
 // Licensed to the NextAudio under one or more agreements.
 // NextAudio licenses this file to you under the MIT license.
 
-using System;
 using System.Threading.Tasks;
+using Discord;
 using Discord.Commands;
 
-namespace DiscordAudioTests
+namespace DiscordAudioTests;
+
+public class AudioModule : ModuleBase<ShardedCommandContext>
 {
-    public class AudioModule : ModuleBase<ShardedCommandContext>
+    private readonly AudioService _audioService;
+
+    public AudioModule(AudioService audioService)
     {
-        private readonly AudioService _audioService;
+        _audioService = audioService;
+    }
 
-        public AudioModule(AudioService audioService)
+    private AudioPlayer Player => _audioService.GetOrCreatePlayerForContext(Context);
+
+    private bool UserConnected => (Context.User as IGuildUser).VoiceChannel != null;
+
+    [Command("play")]
+    public async Task PlayAsync([Remainder] string url)
+    {
+        if (!UserConnected)
         {
-            _audioService = audioService;
+            _ = await ReplyAsync("You're not connected in a voice channel.");
+            return;
         }
 
-        private AudioPlayer Player => _audioService.GetOrCreatePlayerForContext(Context);
+        var directPlayed = Player.Queue.IsEmpty;
 
-        [Command("play")]
-        public Task PlayAsync([Remainder] string url)
+        await _audioService.PlayAsync(Context, url);
+
+        if (directPlayed)
         {
-            return _audioService.PlayAsync(Context, url);
+            return;
         }
 
-        [Command("skip")]
-        public Task SkipAsync()
+        _ = await ReplyAsync($"Song '{url}' added to queue.");
+    }
+
+    [Command("skip")]
+    public async Task SkipAsync()
+    {
+        if (!UserConnected)
         {
-            return Player.SkipAsync().AsTask();
+            _ = await ReplyAsync("You're not connected in a voice channel.");
+            return;
         }
 
-        [Command("pause"), Alias("resume")]
-        public Task PauseOrResumeAsync()
+        if (Player.Queue.IsEmpty)
         {
-            return Player.PauseOrResumeAsync().AsTask();
+            _ = await ReplyAsync("No songs to Skip.");
+            return;
         }
 
-        [Command("seek")]
-        public Task SeekAsync([Remainder] TimeSpan time)
+        await Player.SkipAsync();
+        _ = await ReplyAsync("song skipped.");
+    }
+
+    [Command("pause"), Alias("resume")]
+    public async Task PauseOrResumeAsync()
+    {
+        if (!UserConnected)
         {
-            return Player.SeekAsync(time).AsTask();
+            _ = await ReplyAsync("You're not connected in a voice channel.");
+            return;
         }
 
-        [Command("stop")]
-        public Task StopAsync()
+        var paused = Player.IsPaused;
+
+        await Player.PauseOrResumeAsync();
+
+        if (paused)
         {
-            return _audioService.StopAsync(Context);
+            _ = await ReplyAsync("Player resumed.");
+            return;
         }
+
+        _ = await ReplyAsync("Player paused.");
+    }
+
+    // disabled for now
+    // [Command("seek")]
+    // public Task SeekAsync([Remainder] TimeSpan time)
+    // {
+    //     return Player.SeekAsync(time).AsTask();
+    // }
+
+    [Command("stop")]
+    public async Task StopAsync()
+    {
+        if (!UserConnected)
+        {
+            _ = await ReplyAsync("You're not connected in a voice channel.");
+            return;
+        }
+
+        await _audioService.StopAsync(Context);
+        _ = await ReplyAsync("Player stopped.");
     }
 }
