@@ -1,26 +1,25 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-using Microsoft.Extensions.DependencyInjection;
+// Licensed to the NextAudio under one or more agreements.
+// NextAudio licenses this file to you under the MIT license.
+
 using Microsoft.Extensions.Hosting;
 
-namespace DiscordAudioTests
+namespace DiscordAudioTests;
+
+public class Program
 {
-    public class Program
+    public static void Main(string[] args)
     {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+        CreateHostBuilder(args).Build().Run();
+    }
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureServices((hostContext, services) =>
-                {
-                    var startup = new Startup(hostContext.HostingEnvironment, hostContext.Configuration);
+    public static IHostBuilder CreateHostBuilder(string[] args)
+    {
+        return Host.CreateDefaultBuilder(args)
+            .ConfigureServices((hostContext, services) =>
+            {
+                var startup = new Startup(hostContext.HostingEnvironment, hostContext.Configuration);
 
-                    Startup.ConfigureServices(services);
-                });
+                Startup.ConfigureServices(services);
+            });
     }
 }
